+++ src/components/PortfolioOld.js
@@ -1,5 +1,5 @@
 // Portfolio component - will be displayed when the Portfolio tab in main nav is clicked
-import Card from "./Card";
+import Card from "./CardOld";
 import planet from "../images/planet.gif";
 import planner from "../images/planner-mockup.png";
 import weather from "../images/weather-dash.png";
